pipeline/requirements.txt
altair
boto3
openpyxl
pandas
pyarrow
python-dotenv
redshift_connector
streamlit


